streamingccl: share heartbeatSender between ingestors
This moves the heartbeatSender, mockStreamClient, and
errorStreamClient into the streamclient package for easier sharing
between the logical ingestion job and physical ingestion job.

It minimally integrates the heartbeatSender into the logical
replication job and all heartbeat errors are ignored. A follow-up
commit will add a frontier processor that correctly handles errors
from the heartbeat sender.

Epic: none
Release note: None
